Safari-611.1.10/LayoutTests/fast/media/mq-pointer-matchMedia-expected.txt

Test the pointer media feature with machMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


window.matchMedia("(pointer)").matches = true
window.matchMedia("(pointer: fine)").matches = true
window.matchMedia("(pointer: coarse)").matches = false
window.matchMedia("(pointer: none)").matches = false
window.matchMedia("(pointer: Fine)").matches = true
window.matchMedia("(pointer: Coarse)").matches = false
window.matchMedia("(pointer: None)").matches = false
window.matchMedia("(Pointer: fine)").matches = true
window.matchMedia("(Pointer: coarse)").matches = false
window.matchMedia("(Pointer: none)").matches = false
PASS successfullyParsed is true

TEST COMPLETE
